Item list building and tracker error handling shuffled.
